Main Code
Just import this to test for yourself if you want to or you can copy and paste the scripts and follow along with the tutorial.
Credit to @SlothInTheHat for the Tilemap code but I have improved and revised some bugs that also adds a chunk unloader.

Free top down smooth player movement too included.

Co-Authored-By: SlothInTheHat <71303537+slothinthehat@users.noreply.github.com>